+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.0] - 2022-12-08
+
+- Create Footer and custom UI for sidebar
+- Create notification page
+
 ## [3.3.0] - 2022-12-07
 
 - Check deadline in submit task